+++ test/cram/heavy/klee/ok-checkutf8.t/checkutf8.c
@@ -1,0 +1,1 @@
+../../common-files/checkutf8/checkutf8.c